02/03data_classes/pyproject.toml
[tool.poetry]
name = "data_classes"
version = "0.1.0"
description = "Esplorando data classes"
authors = ["Burlesco70 <mario.nardi.70@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.8.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
